backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
geoalchemy2
shapely
python-dotenv
httpx